+++ client/dep/NclNitroApi/dep/ncl-hl1-source-sdk/vgui2/vgui_controls/MenuItem.cpp
@@ -222,7 +222,8 @@ void MenuItem::OnCursorEntered()
 	// forward the message on to the parent of this menu.
 	KeyValues *msg = new KeyValues ("CursorEnteredMenuItem");
 	// tell the parent this menuitem is the one that was entered so it can highlight it
-	msg->SetInt("VPanel", GetVPanel());
+	// 64-bit: pack VPANEL as uint64 (SetInt truncates → sign-extended garbage → crash).
+	msg->SetUint64("VPanel", static_cast<uint64>(GetVPanel()));
 
 	ivgui()->PostMessage(GetVParent(), msg, NULL);
 }
@@ -236,7 +237,7 @@ void MenuItem::OnCursorExited()
 	// forward the message on to the parent of this menu.
 	KeyValues *msg = new KeyValues ("CursorExitedMenuItem");
 	// tell the parent this menuitem is the one that was entered so it can unhighlight it
-	msg->SetInt("VPanel", GetVPanel());
+	msg->SetUint64("VPanel", static_cast<uint64>(GetVPanel()));
 
 	ivgui()->PostMessage(GetVParent(), msg, NULL);
 }
